Goodbye Redundancy at the End of Game
GENERAL

*Created EndGame.unity Scene and Script.
*Deleted WonGame.unity and LostGame.unity Scenes and Scripts.

SCENES

EndGame.unity
*Carries same structure and function as LostGame.unity and WonGame.unity Scenes.
*Now initializes UI text based on boolean won being true or false.

SCRIPTS

EndGame.cs
*Shows Final Score to UI.
*Shows Won or Lost text to UI.

GameManager.cs
*Now calls upon EndGame.cs method EndGameScene( bool won ).

SceneTransitions.cs
*Deleted LostGameScene() and WonGameScene() methods.
*Added EndGameScene( bool won ) method to transition to EndGame.unity Scene and initialize texts on that Scene based on the boolean won being true or false.
