+++ Artemis Project/Assets/Scenes/EndGame.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e948bc6af11303d4c87646071719ef40
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
