--- Artemis Project/Assets/Scenes/LostGame.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e948bc6af11303d4c87646071719ef40
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
