--- Artemis Project/Assets/Scenes/WonGame.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a80ba9647deb355479cd99e0171ce168
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
