+++ Artemis Project/Assets/Scripts/EndGame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afcc1fb14bfc55e4ca105ab2a3eefe8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
