--- Artemis Project/Assets/Scripts/LostGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6de0503203559bb4fb933f465e427b21
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
