--- Artemis Project/Assets/Scripts/WonGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5cd5fce6f8a03ac40945eb34a7f76341
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
