+++ Artemis Project/Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp
@@ -430,10 +430,10 @@
 "Assets/SaveGame/Scripts/TestSaveGame.cs"
 "Assets/Scripts/BackgroundMove.cs"
 "Assets/Scripts/CreditsScene.cs"
+"Assets/Scripts/EndGame.cs"
 "Assets/Scripts/Fade.cs"
 "Assets/Scripts/FindAndInit.cs"
 "Assets/Scripts/GameManager.cs"
-"Assets/Scripts/LostGame.cs"
 "Assets/Scripts/MenuScene.cs"
 "Assets/Scripts/Play1Scene.cs"
 "Assets/Scripts/Player.cs"
@@ -442,7 +442,6 @@
 "Assets/Scripts/Questions.cs"
 "Assets/Scripts/SceneTransitions.cs"
 "Assets/Scripts/SettingsScene.cs"
-"Assets/Scripts/WonGame.cs"
 "Assets/TextMesh Pro/Examples & Extras/Scripts/Benchmark01.cs"
 "Assets/TextMesh Pro/Examples & Extras/Scripts/Benchmark01_UGUI.cs"
 "Assets/TextMesh Pro/Examples & Extras/Scripts/Benchmark02.cs"
