+++ Artemis Project/Library/StateCache/SceneView/32/32f74c6b2b9b75372837a1c66e6a1f62.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":-1.9796245098114014,"y":-163.84568786621095,"z":-47.783424377441409},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":57.19029998779297,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":34.122779846191409,"y":-148.4432373046875,"z":-47.7576904296875},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":54.616737365722659,"orthographic":true}
