+++ Artemis Project/UserSettings/EditorUserSettings.asset
@@ -12,16 +12,16 @@ EditorUserSettings:
       value: 550103045050580c545a587313755c44444e192c282a27637c791c36e6e4373e
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 0608005e07565f5a0b5f5d7140705b44154f4f7c787e76677c7e4c6ae7e66068
-      flags: 0
-    RecentlyUsedSceneGuid-3:
-      value: 02090404040c5f0f5a0a092043760a44414e1b2e757123607c7e4c30e7b16260
-      flags: 0
-    RecentlyUsedSceneGuid-4:
       value: 5b520d0503545b0d0c0c0a2715770748154f4d2c7d7d7e627a7d4a35b4e1646a
       flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-3:
       value: 05575751040c0b5a555e0d7a417759444f1649282d7970692e7a1e64b2b2303e
+      flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 02090404040c5f0f5a0a092043760a44414e1b2e757123607c7e4c30e7b16260
+      flags: 0
+    RecentlyUsedSceneGuid-5:
+      value: 0608005e07565f5a0b5f5d7140705b44154f4f7c787e76677c7e4c6ae7e66068
       flags: 0
     RecentlyUsedSceneGuid-6:
       value: 5350040353535e035a5c557a167a5d44154f4c2c7f7974622c7a4432b1b4626f
